requirements.txt
openai
langchain
pinecone-client
python-dotenv
tiktoken
docx2txt
PyPDF2
requests
pandas
aiohttp
PyYAML
numpy
langchain_openai
langchain-openai
langchain-community
langchain-pinecone
python-dotenv